Added a single line example for left/right case only.
