+++ examples/menu.markup
@@ -1,7 +1,7 @@
 <view name="main">
     Welcome {{player.name}} to UIL!
     <group class="btns-group">
-        <button class="btn-play" goto-view="play">Play</button>
+        <button class="btn-one-line" goto-view="play">Play</button>
         <button class="btn-one-line" action="quit">Options</button>
         <button class="btn" action="quit">Foo</button>
         <button class="btn" action="quit">Bar</button>
